Modularize mouse input module.
